refactor: Restructure app initialization and middleware setup
- Moved environment variable checks, middleware initialization, error handling, and route setup into separate modules for better organization and maintainability.
- Added process handlers for graceful shutdown and error handling.
- Updated database connection options to include retry settings.
- Refactored authentication service error messages for clarity.
