+++ backend/src/config/database.js
@@ -7,10 +7,11 @@ class Database {
         }
 
         const options = {
-            useNewUrlParser: true,
-            useUnifiedTopology: true,
             serverSelectionTimeoutMS: 5000,
             socketTimeoutMS: 45000,
+            retryWrites: true,
+            w: 'majority',
+            retryReads: true,
         };
 
         try {
